Fix versions according to Gemfile.lock
Considering so much time has passed since this project was last updated,
I think it is best to tie its dependencies to all the old versions that
show up in Gemfile.lock.

I have to assume that this project will work with those old dependencies.
I don't want to go down the rabbit hole of updating all of the deps,
especially because I'm not familiar with middleman or this project.